requirements.txt
a2a-sdk
a2a-sdk[all]
google-adk[a2a]
google-auth
google-cloud-secret-manager
httpx
tenacity
pandas
numpy
pyarrow
duckdb
scikit-learn
lightgbm
xgboost
geopandas
shapely
pyproj
rasterio
pydantic
structlog
uvicorn
fastapi
pytest
pytest-asyncio
requests-mockjoblib
